Outcome needs to be a slice.
